+++ src/app/my/recent-success/recent-success-list/recent-success-list.module.scss
