+++ src/app/my/success/page.tsx
@@ -1,8 +1,8 @@
 'use client';
 import React from 'react';
 
+import RecentSuccessList from '@/app/my/recent-success/recent-success-list/recent-success-list';
 import SuccessList from '@/app/my/success/_shared/components/success-list';
-import RecentSuccessList from '@/shared/components/features/recent-success/recent-success-list/recent-success-list';
 import PageFrame from '@/shared/components/layout/page-frame/page-frame';
 import Spacer from '@/shared/components/ui/spacer/spacer';
 import { useMember } from '@/shared/context/member-context';
