+++ src/shared/components/features/draggable-card-list/draggable-card-list.module.scss
